Changes to better handle poller timeouts and the C glue code.
The flags parameter to send / receive were assigned to a long
variable, which was received in the C glue code as a jlong,
which was finally cast to an int when being passed into
zmq_send() and zmq_recv(). I changed them into int / jint and
got rid of the cast.

I have implemented the timeout handling in the poll() methods:

poll() will behave depending on whether setTimeout() has been
called: if yes, it will use that timeout value; if not, it
will block indefinitely.

poll(timeout) will always honor whatever timeout value it is
explicilty receiving.

[gs]etTimeOut() methods have been deprecated.