fixed invalid argument() exception for setSendBufferSize()
With ZMQ version 3.0 the option value type of ZMQ_SNDBUF and ZMQ_RCVBUF
changed from uint64_t to int. That type change did not go into jzmq lib
that's why the mentioned exception is thrown when calling
setSendBufferSize() which calls setLongSockopt() in the lib code.
I've added the ZMQ_SNDBUF and ZMQ_RCVBUF defines to the case where the
optval is casted to a int value before zmq_setsockopt() is called
This change is valid for ZMQ versions >= 3.0
jira issue LIBZMQ-436