Problem: Project does not have jenv integration Solution: Add .java-version file for java 1.7
jenv (http://www.jenv.be/) provides a convenient way to manage different
local JDKs. When configured it sets the java version depending on the
project's java configuration, so you don't have to remember what version
to use or to create homegrown bash aliases, etc.

This adds a .java-version file (similar to rbenv's .ruby-version or
pyenv's .python-version) to the project, using Java 1.7 based on the
java version from pom.xml.