Problem: Single-module pom is not sufficient for Java ecosystem cleanup
Solution: Create maven multi-module project while retaining compatibilty

This creates a new maven project with the following structure:

- jzmq-parent (meta project)
- jzmq-core (currently empty, will provide a place for CZMQ code in the future)
- jzmq-jni (Access to low-level Java API via JNI)
- jzmq (compatibility project with dependencies on core and jni)

The existing code is moved to the jzmq-jni project.

The poms are restructured with the following goals:

1) Parent pom.xml is the canonical source of versions for _all_ plugins
and dependencies using maven's dependencyManagement resp.
pluginManagement facility. Because the plugin and depenency versions are
only defined in one place, version properties become redundant and are
thus removed.

2) All JNI-specific pom parts are moved to the jzmq-jni pom, while all
generic pom parts that could also be used for other packages stay in the
parent pom.