Problem: jzmq-jni maven dependency does not consider java.library.path
Hm, so this is where the fun begins.

Everything compiles happily but the tests in jzmq-core fail because
libjzmq is not available. In jzmq-jni everything works because the
java.library.path is configured explicitly in surefire plugin (which is
not very elegant but the only way that works).

Solution: Move affected maven config back to jzmq-parent pom.xml

Actually, only the native.library-path property is required for this to
work but there are three maven profiles that set native.library-path,
native.path and native.os and splitting this up feels unnecessary.

The surefire config must be used in all modules for now so it also moves
back.